Remove geometry CI blockers without widening budgets
The shared fresh-window geometry logic now has explicit owners instead of living inside AppDelegate and SessionPersistence. Geometry resolution, shared hint persistence, restore policy checks, and hint-focused tests are split into narrow files so the PR addresses review feedback while keeping the existing runtime behavior.

Constraint: workflow guard and review feedback objected to file growth in AppDelegate.swift, SessionPersistence.swift, and SessionPersistenceTests.swift

Rejected: Raise .github/swift-file-length-budget.tsv | the feature code can be split by responsibility instead

Rejected: Keep startup geometry helpers on AppDelegate | they are pure frame resolution and do not require AppKit lifecycle state

Confidence: medium

Scope-risk: moderate

Directive: Keep WindowGeometryResolver free of AppDelegate state; AppDelegate should only collect screen/window inputs and apply the returned frame

Tested: git diff --check; plutil -lint GhosttyTabs.xcodeproj/project.pbxproj; workflow-guard-tests passed on prior pushed merge run

Not-tested: Local xcodebuild or test suite; user explicitly prohibited bare xcodebuild and project policy routes tests through CI
