+++ Sources/SessionPersistence.swift
@@ -87,57 +87,6 @@ enum SessionPersistencePolicy {
     }
 }
 
-enum SessionRestorePolicy {
-    static func isRunningUnderAutomatedTests(
-        environment: [String: String] = ProcessInfo.processInfo.environment
-    ) -> Bool {
-        if environment["CMUX_UI_TEST_MODE"] == "1" {
-            return true
-        }
-        if environment.keys.contains(where: { $0.hasPrefix("CMUX_UI_TEST_") }) {
-            return true
-        }
-        if environment["XCTestConfigurationFilePath"] != nil {
-            return true
-        }
-        if environment["XCTestBundlePath"] != nil {
-            return true
-        }
-        if environment["XCTestSessionIdentifier"] != nil {
-            return true
-        }
-        if environment["XCInjectBundle"] != nil {
-            return true
-        }
-        if environment["XCInjectBundleInto"] != nil {
-            return true
-        }
-        if environment["DYLD_INSERT_LIBRARIES"]?.contains("libXCTest") == true {
-            return true
-        }
-        return false
-    }
-
-    static func shouldAttemptRestore(
-        arguments: [String] = CommandLine.arguments,
-        environment: [String: String] = ProcessInfo.processInfo.environment
-    ) -> Bool {
-        if environment["CMUX_DISABLE_SESSION_RESTORE"] == "1" {
-            return false
-        }
-        if isRunningUnderAutomatedTests(environment: environment) {
-            return false
-        }
-
-        let extraArgs = arguments
-            .dropFirst()
-            .filter { !$0.hasPrefix("-psn_") }
-
-        // Any explicit launch argument is treated as an explicit open intent.
-        return extraArgs.isEmpty
-    }
-}
-
 struct SessionRectSnapshot: Codable, Equatable, Sendable {
     let x: Double
     let y: Double
@@ -167,14 +116,6 @@ struct SessionDisplaySnapshot: Codable, Sendable {
     var displayID: UInt32?
     var frame: SessionRectSnapshot?
     var visibleFrame: SessionRectSnapshot?
-}
-
-struct SharedWindowGeometryHint: Codable, Sendable {
-    let version: Int
-    let updatedAt: TimeInterval
-    let writerBundleIdentifier: String
-    let frame: SessionRectSnapshot
-    let display: SessionDisplaySnapshot?
 }
 
 enum SessionSidebarSelection: String, Codable, Sendable, Equatable {
@@ -376,74 +317,6 @@ struct AppSessionSnapshot: Codable, Sendable {
     var version: Int
     var createdAt: TimeInterval
     var windows: [SessionWindowSnapshot]
-}
-
-enum SharedWindowGeometryHintStore {
-    static let schemaVersion = 1
-    private static let defaultBundleIdentifier = "com.cmuxterm.app"
-
-    static func load(fileURL: URL? = nil) -> SharedWindowGeometryHint? {
-        guard let fileURL = fileURL ?? defaultFileURL() else { return nil }
-        guard let data = try? Data(contentsOf: fileURL), !data.isEmpty else { return nil }
-        guard let hint = try? JSONDecoder().decode(SharedWindowGeometryHint.self, from: data) else { return nil }
-        guard hint.version == schemaVersion else { return nil }
-        return hint
-    }
-
-    @discardableResult
-    static func save(
-        from snapshot: AppSessionSnapshot,
-        fileURL: URL? = nil,
-        writerBundleIdentifier: String? = Bundle.main.bundleIdentifier,
-        updatedAt: TimeInterval = Date().timeIntervalSince1970
-    ) -> Bool {
-        guard let primaryWindow = snapshot.windows.first else { return false }
-        guard let frame = primaryWindow.frame else { return false }
-        let hint = SharedWindowGeometryHint(
-            version: schemaVersion,
-            updatedAt: updatedAt,
-            writerBundleIdentifier: normalizedBundleIdentifier(writerBundleIdentifier),
-            frame: frame,
-            display: primaryWindow.display
-        )
-        return save(hint, fileURL: fileURL)
-    }
-
-    @discardableResult
-    static func save(_ hint: SharedWindowGeometryHint, fileURL: URL? = nil) -> Bool {
-        guard let fileURL = fileURL ?? defaultFileURL() else { return false }
-        let directory = fileURL.deletingLastPathComponent()
-        do {
-            try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true, attributes: nil)
-            let encoder = JSONEncoder()
-            encoder.outputFormatting = [.sortedKeys]
-            let data = try encoder.encode(hint)
-            try data.write(to: fileURL, options: .atomic)
-            return true
-        } catch {
-            return false
-        }
-    }
-
-    static func defaultFileURL(appSupportDirectory: URL? = nil) -> URL? {
-        let resolvedAppSupport: URL
-        if let appSupportDirectory {
-            resolvedAppSupport = appSupportDirectory
-        } else if let discovered = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first {
-            resolvedAppSupport = discovered
-        } else {
-            return nil
-        }
-
-        return resolvedAppSupport
-            .appendingPathComponent("cmux", isDirectory: true)
-            .appendingPathComponent("window-geometry-hint.json", isDirectory: false)
-    }
-
-    private static func normalizedBundleIdentifier(_ bundleIdentifier: String?) -> String {
-        let trimmed = bundleIdentifier?.trimmingCharacters(in: .whitespacesAndNewlines)
-        return trimmed?.isEmpty == false ? trimmed! : defaultBundleIdentifier
-    }
 }
 
 enum SessionPersistenceStore {
