+++ Sources/SessionRestorePolicy.swift
@@ -1,0 +1,51 @@
+import Foundation
+
+enum SessionRestorePolicy {
+    static func isRunningUnderAutomatedTests(
+        environment: [String: String] = ProcessInfo.processInfo.environment
+    ) -> Bool {
+        if environment["CMUX_UI_TEST_MODE"] == "1" {
+            return true
+        }
+        if environment.keys.contains(where: { $0.hasPrefix("CMUX_UI_TEST_") }) {
+            return true
+        }
+        if environment["XCTestConfigurationFilePath"] != nil {
+            return true
+        }
+        if environment["XCTestBundlePath"] != nil {
+            return true
+        }
+        if environment["XCTestSessionIdentifier"] != nil {
+            return true
+        }
+        if environment["XCInjectBundle"] != nil {
+            return true
+        }
+        if environment["XCInjectBundleInto"] != nil {
+            return true
+        }
+        if environment["DYLD_INSERT_LIBRARIES"]?.contains("libXCTest") == true {
+            return true
+        }
+        return false
+    }
+
+    static func shouldAttemptRestore(
+        arguments: [String] = CommandLine.arguments,
+        environment: [String: String] = ProcessInfo.processInfo.environment
+    ) -> Bool {
+        if environment["CMUX_DISABLE_SESSION_RESTORE"] == "1" {
+            return false
+        }
+        if isRunningUnderAutomatedTests(environment: environment) {
+            return false
+        }
+
+        let extraArgs = arguments
+            .dropFirst()
+            .filter { !$0.hasPrefix("-psn_") }
+
+        return extraArgs.isEmpty
+    }
+}
